moves example to new file
