+++ src/components/accordion/_test_examples.js
@@ -1,0 +1,15 @@
+export { EXAMPLE_ACCORDION };
+
+const EXAMPLE_ACCORDION = {
+    id: 'accordion-identifier',
+    itemsList: [
+        {
+            title: 'Title for item 1',
+            content: 'Content for item 1',
+        },
+        {
+            title: 'Title for item 2',
+            content: 'Content for item 2',
+        },
+    ],
+};
